move SpeakerPage to pages/ folder
